hard link limits cross filesystem ingest
When hg-tmp is on a different filesystem from the ingest file, os.link() fails with: `OSError: [Errno 18] Invalid cross-device link:`

This proposed change uses shutil.copy2(), which copies the file, preserving metadata, and is able to handle the cross-filesystem case. 

This does change the behavior of the program though. where instead of a hard link it's just a copy. This feels safer, but is wasteful because now there is a copy of the file in `hg-tmp`, `hg-data` and the source location.  

Related to an maybe a potential solution to #3